+++ build_x64/debug/moc_frm_principale.cpp
@@ -22,8 +22,8 @@ QT_BEGIN_MOC_NAMESPACE
 QT_WARNING_PUSH
 QT_WARNING_DISABLE_DEPRECATED
 struct qt_meta_stringdata_FRM_Principale_t {
-    QByteArrayData data[54];
-    char stringdata0[824];
+    QByteArrayData data[55];
+    char stringdata0[845];
 };
 #define QT_MOC_LITERAL(idx, ofs, len) \
     Q_STATIC_BYTE_ARRAY_DATA_HEADER_INITIALIZER_WITH_OFFSET(len, \
@@ -55,37 +55,38 @@ QT_MOC_LITERAL(19, 275, 18), // "comparerLesProgres"
 QT_MOC_LITERAL(20, 294, 18), // "filtreTableOrigine"
 QT_MOC_LITERAL(21, 313, 6), // "filtre"
 QT_MOC_LITERAL(22, 320, 16), // "filtreTableTitre"
-QT_MOC_LITERAL(23, 337, 23), // "filtreTableProgresFinis"
-QT_MOC_LITERAL(24, 361, 24), // "filtreTableConditionFait"
-QT_MOC_LITERAL(25, 386, 24), // "filtreTableTypeCondition"
-QT_MOC_LITERAL(26, 411, 17), // "dateFilterChanged"
-QT_MOC_LITERAL(27, 429, 17), // "effacerLesFiltres"
-QT_MOC_LITERAL(28, 447, 25), // "effacerFiltresSurLesDates"
-QT_MOC_LITERAL(29, 473, 18), // "etatAutoCompletion"
-QT_MOC_LITERAL(30, 492, 4), // "etat"
-QT_MOC_LITERAL(31, 497, 16), // "dataSelectionnee"
-QT_MOC_LITERAL(32, 514, 11), // "QModelIndex"
-QT_MOC_LITERAL(33, 526, 19), // "customMenuRequested"
-QT_MOC_LITERAL(34, 546, 3), // "pos"
-QT_MOC_LITERAL(35, 550, 14), // "masquerProgres"
-QT_MOC_LITERAL(36, 565, 7), // "booleen"
-QT_MOC_LITERAL(37, 573, 16), // "restaurerProgres"
-QT_MOC_LITERAL(38, 590, 13), // "imprimerTable"
-QT_MOC_LITERAL(39, 604, 14), // "graphiqueLigne"
-QT_MOC_LITERAL(40, 619, 15), // "graphiqueSpline"
-QT_MOC_LITERAL(41, 635, 23), // "ouvrirFenGraphiqueStats"
-QT_MOC_LITERAL(42, 659, 7), // "clicked"
-QT_MOC_LITERAL(43, 667, 22), // "fenGraphiqueStatsClose"
-QT_MOC_LITERAL(44, 690, 6), // "result"
-QT_MOC_LITERAL(45, 697, 13), // "ouvrirAPropos"
-QT_MOC_LITERAL(46, 711, 16), // "ouvrirParametres"
-QT_MOC_LITERAL(47, 728, 17), // "verifierMiseAJour"
-QT_MOC_LITERAL(48, 746, 17), // "fichierTelecharge"
-QT_MOC_LITERAL(49, 764, 14), // "QNetworkReply*"
-QT_MOC_LITERAL(50, 779, 6), // "pReply"
-QT_MOC_LITERAL(51, 786, 18), // "comparaisonVersion"
-QT_MOC_LITERAL(52, 805, 13), // "ecrireFichier"
-QT_MOC_LITERAL(53, 819, 4) // "TEST"
+QT_MOC_LITERAL(23, 337, 20), // "filtreTableCategorie"
+QT_MOC_LITERAL(24, 358, 23), // "filtreTableProgresFinis"
+QT_MOC_LITERAL(25, 382, 24), // "filtreTableConditionFait"
+QT_MOC_LITERAL(26, 407, 24), // "filtreTableTypeCondition"
+QT_MOC_LITERAL(27, 432, 17), // "dateFilterChanged"
+QT_MOC_LITERAL(28, 450, 17), // "effacerLesFiltres"
+QT_MOC_LITERAL(29, 468, 25), // "effacerFiltresSurLesDates"
+QT_MOC_LITERAL(30, 494, 18), // "etatAutoCompletion"
+QT_MOC_LITERAL(31, 513, 4), // "etat"
+QT_MOC_LITERAL(32, 518, 16), // "dataSelectionnee"
+QT_MOC_LITERAL(33, 535, 11), // "QModelIndex"
+QT_MOC_LITERAL(34, 547, 19), // "customMenuRequested"
+QT_MOC_LITERAL(35, 567, 3), // "pos"
+QT_MOC_LITERAL(36, 571, 14), // "masquerProgres"
+QT_MOC_LITERAL(37, 586, 7), // "booleen"
+QT_MOC_LITERAL(38, 594, 16), // "restaurerProgres"
+QT_MOC_LITERAL(39, 611, 13), // "imprimerTable"
+QT_MOC_LITERAL(40, 625, 14), // "graphiqueLigne"
+QT_MOC_LITERAL(41, 640, 15), // "graphiqueSpline"
+QT_MOC_LITERAL(42, 656, 23), // "ouvrirFenGraphiqueStats"
+QT_MOC_LITERAL(43, 680, 7), // "clicked"
+QT_MOC_LITERAL(44, 688, 22), // "fenGraphiqueStatsClose"
+QT_MOC_LITERAL(45, 711, 6), // "result"
+QT_MOC_LITERAL(46, 718, 13), // "ouvrirAPropos"
+QT_MOC_LITERAL(47, 732, 16), // "ouvrirParametres"
+QT_MOC_LITERAL(48, 749, 17), // "verifierMiseAJour"
+QT_MOC_LITERAL(49, 767, 17), // "fichierTelecharge"
+QT_MOC_LITERAL(50, 785, 14), // "QNetworkReply*"
+QT_MOC_LITERAL(51, 800, 6), // "pReply"
+QT_MOC_LITERAL(52, 807, 18), // "comparaisonVersion"
+QT_MOC_LITERAL(53, 826, 13), // "ecrireFichier"
+QT_MOC_LITERAL(54, 840, 4) // "TEST"
 
     },
     "FRM_Principale\0downloaded\0\0"
@@ -97,7 +98,7 @@ QT_MOC_LITERAL(53, 819, 4) // "TEST"
     "importProgresBlazeandcave\0importProgresPerso\0"
     "exclureStats\0statut\0comparerLesProgres\0"
     "filtreTableOrigine\0filtre\0filtreTableTitre\0"
-    "filtreTableProgresFinis\0"
+    "filtreTableCategorie\0filtreTableProgresFinis\0"
     "filtreTableConditionFait\0"
     "filtreTableTypeCondition\0dateFilterChanged\0"
     "effacerLesFiltres\0effacerFiltresSurLesDates\0"
@@ -120,7 +121,7 @@ static const uint qt_meta_data_FRM_Principale[] = {
        8,       // revision
        0,       // classname
        0,    0, // classinfo
-      36,   14, // methods
+      37,   14, // methods
        0,    0, // properties
        0,    0, // enums/sets
        0,    0, // constructors
@@ -128,44 +129,45 @@ static const uint qt_meta_data_FRM_Principale[] = {
        2,       // signalCount
 
  // signals: name, argc, parameters, tag, flags
-       1,    1,  194,    2, 0x06 /* Public */,
-       4,    0,  197,    2, 0x06 /* Public */,
+       1,    1,  199,    2, 0x06 /* Public */,
+       4,    0,  202,    2, 0x06 /* Public */,
 
  // slots: name, argc, parameters, tag, flags
-       5,    1,  198,    2, 0x08 /* Private */,
-       7,    1,  201,    2, 0x08 /* Private */,
-       9,    1,  204,    2, 0x08 /* Private */,
-      11,    1,  207,    2, 0x08 /* Private */,
-      13,    1,  210,    2, 0x08 /* Private */,
-      14,    1,  213,    2, 0x08 /* Private */,
-      15,    1,  216,    2, 0x08 /* Private */,
-      16,    1,  219,    2, 0x08 /* Private */,
-      17,    1,  222,    2, 0x08 /* Private */,
-      19,    1,  225,    2, 0x08 /* Private */,
-      20,    1,  228,    2, 0x08 /* Private */,
-      22,    1,  231,    2, 0x08 /* Private */,
-      23,    1,  234,    2, 0x08 /* Private */,
-      24,    1,  237,    2, 0x08 /* Private */,
-      25,    1,  240,    2, 0x08 /* Private */,
-      26,    0,  243,    2, 0x08 /* Private */,
-      27,    1,  244,    2, 0x08 /* Private */,
-      28,    1,  247,    2, 0x08 /* Private */,
-      29,    1,  250,    2, 0x08 /* Private */,
-      31,    1,  253,    2, 0x08 /* Private */,
-      33,    1,  256,    2, 0x08 /* Private */,
-      35,    1,  259,    2, 0x08 /* Private */,
-      37,    1,  262,    2, 0x08 /* Private */,
-      38,    1,  265,    2, 0x08 /* Private */,
-      39,    1,  268,    2, 0x08 /* Private */,
-      40,    1,  271,    2, 0x08 /* Private */,
-      41,    1,  274,    2, 0x08 /* Private */,
-      43,    1,  277,    2, 0x08 /* Private */,
-      45,    0,  280,    2, 0x08 /* Private */,
-      46,    0,  281,    2, 0x08 /* Private */,
-      47,    0,  282,    2, 0x08 /* Private */,
-      48,    1,  283,    2, 0x08 /* Private */,
-      51,    1,  286,    2, 0x08 /* Private */,
-      53,    1,  289,    2, 0x08 /* Private */,
+       5,    1,  203,    2, 0x08 /* Private */,
+       7,    1,  206,    2, 0x08 /* Private */,
+       9,    1,  209,    2, 0x08 /* Private */,
+      11,    1,  212,    2, 0x08 /* Private */,
+      13,    1,  215,    2, 0x08 /* Private */,
+      14,    1,  218,    2, 0x08 /* Private */,
+      15,    1,  221,    2, 0x08 /* Private */,
+      16,    1,  224,    2, 0x08 /* Private */,
+      17,    1,  227,    2, 0x08 /* Private */,
+      19,    1,  230,    2, 0x08 /* Private */,
+      20,    1,  233,    2, 0x08 /* Private */,
+      22,    1,  236,    2, 0x08 /* Private */,
+      23,    1,  239,    2, 0x08 /* Private */,
+      24,    1,  242,    2, 0x08 /* Private */,
+      25,    1,  245,    2, 0x08 /* Private */,
+      26,    1,  248,    2, 0x08 /* Private */,
+      27,    0,  251,    2, 0x08 /* Private */,
+      28,    1,  252,    2, 0x08 /* Private */,
+      29,    1,  255,    2, 0x08 /* Private */,
+      30,    1,  258,    2, 0x08 /* Private */,
+      32,    1,  261,    2, 0x08 /* Private */,
+      34,    1,  264,    2, 0x08 /* Private */,
+      36,    1,  267,    2, 0x08 /* Private */,
+      38,    1,  270,    2, 0x08 /* Private */,
+      39,    1,  273,    2, 0x08 /* Private */,
+      40,    1,  276,    2, 0x08 /* Private */,
+      41,    1,  279,    2, 0x08 /* Private */,
+      42,    1,  282,    2, 0x08 /* Private */,
+      44,    1,  285,    2, 0x08 /* Private */,
+      46,    0,  288,    2, 0x08 /* Private */,
+      47,    0,  289,    2, 0x08 /* Private */,
+      48,    0,  290,    2, 0x08 /* Private */,
+      49,    1,  291,    2, 0x08 /* Private */,
+      52,    1,  294,    2, 0x08 /* Private */,
+      54,    1,  297,    2, 0x08 /* Private */,
 
  // signals: parameters
     QMetaType::Void, QMetaType::Bool,    3,
@@ -187,24 +189,25 @@ static const uint qt_meta_data_FRM_Principale[] = {
     QMetaType::Void, QMetaType::QString,   21,
     QMetaType::Void, QMetaType::QString,   21,
     QMetaType::Void, QMetaType::QString,   21,
+    QMetaType::Void, QMetaType::QString,   21,
     QMetaType::Void,
     QMetaType::Void, QMetaType::Bool,   12,
     QMetaType::Void, QMetaType::Bool,   12,
-    QMetaType::Void, QMetaType::Int,   30,
-    QMetaType::Void, 0x80000000 | 32,    8,
-    QMetaType::Void, QMetaType::QPoint,   34,
-    QMetaType::Void, QMetaType::Bool,   36,
-    QMetaType::Void, QMetaType::Bool,   36,
+    QMetaType::Void, QMetaType::Int,   31,
+    QMetaType::Void, 0x80000000 | 33,    8,
+    QMetaType::Void, QMetaType::QPoint,   35,
+    QMetaType::Void, QMetaType::Bool,   37,
+    QMetaType::Void, QMetaType::Bool,   37,
     QMetaType::Void, QMetaType::Bool,   12,
     QMetaType::Void, QMetaType::Bool,   12,
     QMetaType::Void, QMetaType::Bool,   12,
-    QMetaType::Void, QMetaType::Bool,   42,
-    QMetaType::Void, QMetaType::Int,   44,
+    QMetaType::Void, QMetaType::Bool,   43,
+    QMetaType::Void, QMetaType::Int,   45,
     QMetaType::Void,
     QMetaType::Void,
     QMetaType::Void,
-    QMetaType::Void, 0x80000000 | 49,   50,
-    QMetaType::Void, QMetaType::Bool,   52,
+    QMetaType::Void, 0x80000000 | 50,   51,
+    QMetaType::Void, QMetaType::Bool,   53,
     QMetaType::Void, QMetaType::Bool,   12,
 
        0        // eod
@@ -230,34 +233,35 @@ void FRM_Principale::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _
         case 11: _t->comparerLesProgres((*reinterpret_cast< bool(*)>(_a[1]))); break;
         case 12: _t->filtreTableOrigine((*reinterpret_cast< QString(*)>(_a[1]))); break;
         case 13: _t->filtreTableTitre((*reinterpret_cast< QString(*)>(_a[1]))); break;
-        case 14: _t->filtreTableProgresFinis((*reinterpret_cast< QString(*)>(_a[1]))); break;
-        case 15: _t->filtreTableConditionFait((*reinterpret_cast< QString(*)>(_a[1]))); break;
-        case 16: _t->filtreTableTypeCondition((*reinterpret_cast< QString(*)>(_a[1]))); break;
-        case 17: _t->dateFilterChanged(); break;
-        case 18: _t->effacerLesFiltres((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 19: _t->effacerFiltresSurLesDates((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 20: _t->etatAutoCompletion((*reinterpret_cast< int(*)>(_a[1]))); break;
-        case 21: _t->dataSelectionnee((*reinterpret_cast< const QModelIndex(*)>(_a[1]))); break;
-        case 22: _t->customMenuRequested((*reinterpret_cast< QPoint(*)>(_a[1]))); break;
-        case 23: _t->masquerProgres((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 24: _t->restaurerProgres((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 25: _t->imprimerTable((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 26: _t->graphiqueLigne((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 27: _t->graphiqueSpline((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 28: _t->ouvrirFenGraphiqueStats((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 29: _t->fenGraphiqueStatsClose((*reinterpret_cast< int(*)>(_a[1]))); break;
-        case 30: _t->ouvrirAPropos(); break;
-        case 31: _t->ouvrirParametres(); break;
-        case 32: _t->verifierMiseAJour(); break;
-        case 33: _t->fichierTelecharge((*reinterpret_cast< QNetworkReply*(*)>(_a[1]))); break;
-        case 34: _t->comparaisonVersion((*reinterpret_cast< bool(*)>(_a[1]))); break;
-        case 35: _t->TEST((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 14: _t->filtreTableCategorie((*reinterpret_cast< QString(*)>(_a[1]))); break;
+        case 15: _t->filtreTableProgresFinis((*reinterpret_cast< QString(*)>(_a[1]))); break;
+        case 16: _t->filtreTableConditionFait((*reinterpret_cast< QString(*)>(_a[1]))); break;
+        case 17: _t->filtreTableTypeCondition((*reinterpret_cast< QString(*)>(_a[1]))); break;
+        case 18: _t->dateFilterChanged(); break;
+        case 19: _t->effacerLesFiltres((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 20: _t->effacerFiltresSurLesDates((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 21: _t->etatAutoCompletion((*reinterpret_cast< int(*)>(_a[1]))); break;
+        case 22: _t->dataSelectionnee((*reinterpret_cast< const QModelIndex(*)>(_a[1]))); break;
+        case 23: _t->customMenuRequested((*reinterpret_cast< QPoint(*)>(_a[1]))); break;
+        case 24: _t->masquerProgres((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 25: _t->restaurerProgres((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 26: _t->imprimerTable((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 27: _t->graphiqueLigne((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 28: _t->graphiqueSpline((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 29: _t->ouvrirFenGraphiqueStats((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 30: _t->fenGraphiqueStatsClose((*reinterpret_cast< int(*)>(_a[1]))); break;
+        case 31: _t->ouvrirAPropos(); break;
+        case 32: _t->ouvrirParametres(); break;
+        case 33: _t->verifierMiseAJour(); break;
+        case 34: _t->fichierTelecharge((*reinterpret_cast< QNetworkReply*(*)>(_a[1]))); break;
+        case 35: _t->comparaisonVersion((*reinterpret_cast< bool(*)>(_a[1]))); break;
+        case 36: _t->TEST((*reinterpret_cast< bool(*)>(_a[1]))); break;
         default: ;
         }
     } else if (_c == QMetaObject::RegisterMethodArgumentMetaType) {
         switch (_id) {
         default: *reinterpret_cast<int*>(_a[0]) = -1; break;
-        case 33:
+        case 34:
             switch (*reinterpret_cast<int*>(_a[1])) {
             default: *reinterpret_cast<int*>(_a[0]) = -1; break;
             case 0:
@@ -313,13 +317,13 @@ int FRM_Principale::qt_metacall(QMetaObject::Call _c, int _id, void **_a)
     if (_id < 0)
         return _id;
     if (_c == QMetaObject::InvokeMetaMethod) {
-        if (_id < 36)
+        if (_id < 37)
             qt_static_metacall(this, _c, _id, _a);
-        _id -= 36;
+        _id -= 37;
     } else if (_c == QMetaObject::RegisterMethodArgumentMetaType) {
-        if (_id < 36)
+        if (_id < 37)
             qt_static_metacall(this, _c, _id, _a);
-        _id -= 36;
+        _id -= 37;
     }
     return _id;
 }